requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field[phonenumberslite]
django-defender
redis[hiredis]